test: all tests running
Modifications to positionwise_ffn.py: corrected the calculations while non_linear_proj is not None

+++ transformerx/layers/dot_product_attention.py
@@ -131,7 +131,7 @@ class DotProductAttention(tf.keras.layers.Layer):
         if self.scaled:
             depth = queries.shape[-1]
 
-            scores = scores / tf.math.sqrt(tf.cast(depth, dtype=tf.float32))
+            scores = scores / tf.math.sqrt(tf.cast(depth, dtype=queries.dtype))
 
         # apply causal mask
         if self.causal_mask:
